Errors fixed in python exercises: basic operations, simple_interest, rectangle_area, future_age_calculator

+++ python_introduction/simple_interest.py
@@ -1,9 +1,13 @@
 # Define variables
+
 principal = 1000
 rate = 0.05
 time = 3
 
 # Calculating simple interest
 
-interest = int(principal * rate * time)
-print('The simple interest is:', int(interest))
+interest = principal * rate * time
+
+# Display output
+
+print('The simple interest is:', interest)
